refactor: migrate changed files to @/ path aliases per project conventions
- agents/index.ts: export makeTieredParseRetryStrategy + TieredInspection from barrel
- agents/retry/tiered-parse-retry.ts: replace @/review/truncation (circular dep risk)
  with inline check using MAX_AGENT_OUTPUT_CHARS from ../acp/adapter
- operations/plan-critic-llm.ts: @/agents/retry → @/agents
- runtime/index.ts: export KNOWN_SESSION_ROLES, isSessionRole, SessionRole types
- test/unit/debate/verifiers/checks.test.ts: @/debate/verifiers/checks → @/debate
- test/unit/runtime/session-role-plan-critic.test.ts: @/runtime/session-role → @/runtime
- test/unit/agents/retry/tiered-parse-retry.test.ts: ../../../../src/agents/retry/types → @/agents

check:alias-internals: 0 violations; deep-relatives: ↓8 from baseline
