+++ src/agents/retry/tiered-parse-retry.ts
@@ -1,5 +1,5 @@
-import { getSafeLogger } from "../../logger";
-import { looksLikeTruncatedJson } from "../../review/truncation";
+import { getSafeLogger } from "@/logger";
+import { MAX_AGENT_OUTPUT_CHARS } from "../acp/adapter";
 import { ParseValidationError } from "./types";
 import type { RetryStrategy } from "./types";
 
@@ -34,7 +34,7 @@ export function makeTieredParseRetryStrategy<TOutput, TKind extends string, TPar
         return { retry: false, fallback: opts.exhaustedFallback(inspection, ctx.lastOutput) };
       }
 
-      const isTruncated = looksLikeTruncatedJson(ctx.lastOutput);
+      const isTruncated = ctx.lastOutput.trimEnd().length >= MAX_AGENT_OUTPUT_CHARS - 100;
       const logger = opts._logger ?? getSafeLogger();
       logger?.warn(opts.reviewerKind, `Parse retry — ${inspection.kind ?? "unknown"}`, {
         storyId: ctx.storyId,
