+++ src/runtime/index.ts
@@ -60,6 +60,8 @@ export { AgentStreamEventBus } from "./agent-stream-events";
 export { attachAgentIdleWatchdog, attachAgentStreamLogging } from "./middleware";
 export type { WatchdogState } from "./middleware";
 export { formatSessionName } from "./session-name";
+export { KNOWN_SESSION_ROLES, isSessionRole } from "./session-role";
+export type { SessionRole, CanonicalSessionRole } from "./session-role";
 
 import { basename, join } from "node:path";
 import type { IAgentManager } from "../agents";
